// payload-types.ts
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "projects".
 */
export interface Project {
  id: number;
  title: string;
  projectBehaviour: {
    /**
     * e.g. 'ai-databases'
     */
    slug: string;
    isFeatured?: boolean | null;
    isHighlighted?: boolean | null;
  };
  projectDetails?: {
    featuredImage?: (number | null) | Media;
    subtitle?: string | null;
    description?: string | null;
    overview?: {
      root: {
        type: string;
        children: {
          type: any;
          version: number;
          [k: string]: unknown;
        }[];
        direction: ('ltr' | 'rtl') | null;
        format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
        indent: number;
        version: number;
      };
      [k: string]: unknown;
    } | null;
  };
  externalLinks?:
    | {
        label: string;
        url: string;
        id?: string | null;
      }[]
    | null;
  images?:
    | {
        image: number | Media;
        id?: string | null;
      }[]
    | null;
  tags?:
    | {
        label: string;
        id?: string | null;
      }[]
    | null;
  technologies?:
    | {
        label: string;
        id?: string | null;
      }[]
    | null;
  projectArticle?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  updatedAt: string;
  createdAt: string;
}